+++ src/hangar/backends/specparse.pyx
@@ -4,6 +4,7 @@ from .specs cimport HDF5_01_DataHashSpec, \
     HDF5_00_DataHashSpec, \
     NUMPY_10_DataHashSpec, \
     LMDB_30_DataHashSpec, \
+    LMDB_31_DataHashSpec, \
     REMOTE_50_DataHashSpec
 
 
@@ -155,6 +156,29 @@ cdef LMDB_30_DataHashSpec LMDB_30_Parser(str inp):
     return res
 
 
+cdef LMDB_31_DataHashSpec LMDB_31_Parser(str inp):
+    cdef str fmt, uid, row_idx, checksum
+    cdef unsigned char i, c, cc
+    cdef unsigned char n = len(inp)
+    cdef LMDB_31_DataHashSpec res
+
+    c = 0
+    cc = 0
+    for i in range(n):
+        if inp[i] == ':':
+            if cc == 0:
+                fmt = inp[c:i]
+            elif cc == 1:
+                uid = inp[c:i]
+            elif cc == 2:
+                row_idx = inp[c:i]
+            c = i + 1
+            cc = cc + 1
+    checksum = inp[c:n]
+
+    res = LMDB_31_DataHashSpec(fmt, uid, row_idx, checksum)
+    return res
+
 
 cdef REMOTE_50_DataHashSpec REMOTE_50_Parser(str inp):
     cdef str fmt, schema_hash
@@ -184,6 +208,8 @@ cpdef object backend_decoder(bytes inp):
         return NUMPY_10_Parser(inp_str)
     elif backend == '30':
         return LMDB_30_Parser(inp_str)
+    elif backend == '31':
+        return LMDB_31_Parser(inp_str)
     elif backend == '50':
         return REMOTE_50_Parser(inp_str)
     else: