+++ tests/test_column_definition_permutations.py
@@ -8,7 +8,7 @@ import numpy as np
 
 
 def assert_equal(expected, actual):
-    if isinstance(expected, str):
+    if isinstance(expected, (str, bytes)):
         assert expected == actual
     elif isinstance(expected, np.ndarray):
         assert np.allclose(expected, actual)
@@ -41,6 +41,15 @@ def str_generate_data_variable_shape(
     return res
 
 
+def bytes_generate_data_variable_shape(
+        length=20, *,
+        _ALPHABET=''.join([string.printable, '\x01', '\x12', '\x25', '\x26', '\x27', '\x91'])
+):
+    tokens = [secrets.choice(_ALPHABET) for i in range(length)]
+    res = ''.join(tokens).encode()
+    return res
+
+
 column_settings = {
     'ndarray': {
         'fixed_shape': ['00', '01', '10'],
@@ -48,6 +57,9 @@ column_settings = {
     },
     'str': {
         'variable_shape': ['30']
+    },
+    'bytes': {
+        'variable_shape': ['31']
     }
 }
 
@@ -59,13 +71,17 @@ column_data_generators = {
     },
     'str': {
         'variable_shape': str_generate_data_variable_shape
+    },
+    'bytes': {
+        'variable_shape': bytes_generate_data_variable_shape
     }
 }
 
 
 column_layouts = {
     'ndarray': ['flat', 'nested'],
-    'str': ['flat', 'nested']
+    'str': ['flat', 'nested'],
+    'bytes': ['flat', 'nested']
 }
 
 
@@ -116,11 +132,11 @@ def column_permutation_repo(repo, num_samples_gen, num_subsamples_gen):
 
                     if col_dtype == 'ndarray':
                         col = co.add_ndarray_column(name,
-                                                       shape=shape,
-                                                       dtype=dtype,
-                                                       variable_shape=is_var,
-                                                       contains_subsamples=has_subs,
-                                                       backend=backend)
+                                                    shape=shape,
+                                                    dtype=dtype,
+                                                    variable_shape=is_var,
+                                                    contains_subsamples=has_subs,
+                                                    backend=backend)
                         data_partial = partial(generator, shape, dtype)
                         if layout == 'flat':
                             column_data_copy[name] = add_data_to_column(col, data_partial, nsamp)
@@ -130,6 +146,13 @@ def column_permutation_repo(repo, num_samples_gen, num_subsamples_gen):
                             raise ValueError(f'invalid layout {layout}')
                     elif col_dtype == 'str':
                         col = co.add_str_column(name, contains_subsamples=has_subs, backend=backend)
+                        data_partial = partial(generator)
+                        if layout == 'flat':
+                            column_data_copy[name] = add_data_to_column(col, data_partial, nsamp)
+                        elif layout == 'nested':
+                            column_data_copy[name] = add_data_to_column(col, data_partial, nsamp, nsubs)
+                    elif col_dtype == 'bytes':
+                        col = co.add_bytes_column(name, contains_subsamples=has_subs, backend=backend)
                         data_partial = partial(generator)
                         if layout == 'flat':
                             column_data_copy[name] = add_data_to_column(col, data_partial, nsamp)
@@ -158,6 +181,87 @@ def column_permutations_write_checkout(column_permutation_repo):
     repo, column_data, column_data_partials = column_permutation_repo
     co = repo.checkout(write=True)
     yield co, column_data, column_data_partials
+    co.close()
+
+
+@pytest.mark.parametrize('column_type,column_kwargs', [
+    ('ndarray', {'prototype': np.array([1, 2, 3])}),
+    ('str', {}),
+    ('bytes', {}),
+])
+@pytest.mark.parametrize('contains_subsamples', [True, False])
+def test_cannot_create_column_within_cm(repo, column_type, column_kwargs, contains_subsamples):
+    co = repo.checkout(write=True)
+    with co:
+        with pytest.raises(PermissionError):
+            if column_type == 'ndarray':
+                co.add_ndarray_column(
+                    'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+            elif column_type == 'str':
+                co.add_str_column(
+                    'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+            elif column_type == 'bytes':
+                co.add_bytes_column(
+                    'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+            else:
+                raise ValueError(column_type)
+    co.close()
+
+
+@pytest.mark.parametrize('column_type,column_kwargs', [
+    ('ndarray', {'prototype': np.array([1, 2, 3])}),
+    ('str', {}),
+    ('bytes', {}),
+])
+def test_contains_subsamples_non_bool_value_fails(repo, column_type, column_kwargs):
+    co = repo.checkout(write=True)
+    with pytest.raises(ValueError):
+        if column_type == 'ndarray':
+            co.add_ndarray_column(
+                'testcol', contains_subsamples=None, **column_kwargs)
+        elif column_type == 'str':
+            co.add_str_column(
+                'testcol', contains_subsamples=None, **column_kwargs)
+        elif column_type == 'bytes':
+            co.add_bytes_column(
+                'testcol', contains_subsamples=None, **column_kwargs)
+        else:
+            raise ValueError(column_type)
+    co.close()
+
+
+@pytest.mark.parametrize('column_type,column_kwargs', [
+    ('ndarray', {'prototype': np.array([1, 2, 3])}),
+    ('str', {}),
+    ('bytes', {}),
+])
+@pytest.mark.parametrize('contains_subsamples', [True, False])
+def test_cannot_create_column_name_exists(repo, column_type, column_kwargs, contains_subsamples):
+    co = repo.checkout(write=True)
+
+    # setup so that a column already exists
+    if column_type == 'ndarray':
+        co.add_ndarray_column(
+            'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+    elif column_type == 'str':
+        co.add_str_column(
+            'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+    elif column_type == 'bytes':
+        co.add_bytes_column(
+            'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+
+    with pytest.raises(LookupError):
+        if column_type == 'ndarray':
+            co.add_ndarray_column(
+                'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+        elif column_type == 'str':
+            co.add_str_column(
+                'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+        elif column_type == 'bytes':
+            co.add_bytes_column(
+                'testcol', contains_subsamples=contains_subsamples, **column_kwargs)
+        else:
+            raise ValueError(column_type)
     co.close()